+++ include/et/imaging/texturedescription.h
@@ -24,30 +24,30 @@ namespace et
         bool load(const std::string& name);
         bool preload(const std::string& name, bool fillWithZero);
         
-		vec2i sizeForMipLevel(size_t level)
+		vec2i sizeForMipLevel(uint32_t level)
 		{
 			vec2i result = size;
-			for (size_t i = 0; i < level; ++i)
+			for (uint32_t i = 0; i < level; ++i)
 				result /= 2;
 			return result;
 		}
 
-		size_t dataSizeForMipLevel(size_t level)
+		uint32_t dataSizeForMipLevel(uint32_t level)
 		{
-			size_t actualSize = static_cast<size_t>(sizeForMipLevel(level).square()) * bitsPerPixel / 8;
-			size_t minimumSize = static_cast<size_t>(minimalSizeForCompressedFormat.square()) * bitsPerPixel / 8;
+			uint32_t actualSize = static_cast<uint32_t>(sizeForMipLevel(level).square()) * bitsPerPixel / 8;
+			uint32_t minimumSize = static_cast<uint32_t>(minimalSizeForCompressedFormat.square()) * bitsPerPixel / 8;
 			return compressed ? std::max(minimalDataSize, std::max(minimumSize, actualSize)) : actualSize;
 		}
 
-		size_t dataSizeForAllMipLevels()
+		uint32_t dataSizeForAllMipLevels()
 		{
-			size_t result = 0;
-			for (size_t i = 0; i < mipMapCount; ++i)
+			uint32_t result = 0;
+			for (uint32_t i = 0; i < mipMapCount; ++i)
 				result += dataSizeForMipLevel(i);
 			return result;
 		}
 
-		size_t dataOffsetForLayer(size_t layer, size_t level)
+		uint32_t dataOffsetForLayer(uint32_t layer, uint32_t level)
 		{
 			if (dataLayout == TextureDataLayout::FacesFirst)
 			{
@@ -62,19 +62,19 @@ namespace et
 			}
 		}
 
-		size_t dataOffsetForMipLevel(size_t level, size_t layer)
+		uint32_t dataOffsetForMipLevel(uint32_t level, uint32_t layer)
 		{
-			size_t result = 0;
+			uint32_t result = 0;
 			if (dataLayout == TextureDataLayout::FacesFirst)
 			{
 				result = dataOffsetForLayer(layer, 0);
-				for (size_t i = 0; i < level; ++i)
+				for (uint32_t i = 0; i < level; ++i)
 					result += dataSizeForMipLevel(i);
 			}
 			else if (dataLayout == TextureDataLayout::MipsFirst)
 			{
 				result = dataOffsetForLayer(layer, level);
-				for (size_t l = 0; l < level; ++l)
+				for (uint32_t l = 0; l < level; ++l)
 					result += layersCount * dataSizeForMipLevel(l);
 			}
 			else
@@ -102,7 +102,7 @@ namespace et
 		uint32_t alignment = 1;
 		uint32_t rowSize = 0;
 
-		size_t minimalDataSize = 0;
+		uint32_t minimalDataSize = 0;
 		
 		TextureTarget target = TextureTarget::Texture_2D;
 		TextureFormat internalformat = TextureFormat::Invalid;
