+++ include/et/rendering/interface/vertexbuffer.h
@@ -36,12 +36,12 @@ namespace et
 
 		virtual void bind() = 0;
 
-		virtual void setData(const void* data, size_t dataSize, bool invalidateExistingData) = 0;
-		virtual void setDataWithOffset(const void* data, size_t offset, size_t dataSize) = 0;
+		virtual void setData(const void* data, uint32_t dataSize, bool invalidateExistingData) = 0;
+		virtual void setDataWithOffset(const void* data, uint32_t offset, uint32_t dataSize) = 0;
 
 		virtual uint64_t dataSize() = 0;
 
-		virtual void* map(size_t offset, size_t dataSize, uint32_t options /* see MapBufferOptions */) = 0;
+		virtual void* map(uint32_t offset, uint32_t dataSize, uint32_t options /* see MapBufferOptions */) = 0;
 		virtual bool mapped() const = 0;
 		virtual void unmap() = 0;
